OMNI-15320: Cleanup the Gradle build and configure new artifacts
The build for Windows and Linux will generate the exact same set
of artifacts with all available modes. We split the artifacts in
three sets to manage the usage better. The main libraries including
the fipscanister with the signature hash are packaged in the common
NAR files to help with the builds. The Crypto engines are packaged
in a separate archive because they are not commonly needed for our
purposes. Finally the FIPS build tools including the fips_premain
source and signature files are packaged in a third archive to use
for the Windows Arbiter builds.