INFRA-3672: Add toggle for toolchain on Windows platform
In order to set the toolchain to build openssl on Windows,
util/buildFOM.bat sets the environment by calling vcvarsall.bat for the
corresponding version of Visual Studio.  This change provides control over
the toolchain on Windows with the ext.toolset property.  This property is
used in other third-party library repos to select the toolchain.