INFRA-4296: Added missing symbol versions to support code that uses them
Code compiled on Ubuntu distribuitions expects OpenSSL to have been compiled
with versioned symbols. As a result this caused the replacement of the default
platform OpenSSL to break things when used to link code. Runtime appears to
not suffer from this and that is still to be understood but with this
change both issues are eliminated so the true root cause of this might
be a forgotten mystery.